Use UpperCamelCase instead of snake_case for namespaces
This makes Foo::bar work for both a namespace Foo and a class Foo.
