+++ CONTRIBUTING.md
@@ -64,9 +64,9 @@ can install the [clang-format](https://clang.llvm.org/docs/ClangFormat.html)
 
 Regarding naming:
 
-* Use UpperCamelCase for types (e.g. classes and structs).
-* Use snake_case for namespaces, functions and variables.
+* Use UpperCamelCase for types (e.g. classes and structs) and namespaces.
+* Use UPPER_CASE names for macros.
+* Use snake_case for other names (functions, variables, enum values, etc.).
 * Use an "m_" prefix for non-public member variables.
 * Use a "g_" prefix for global mutable variables.
 * Use a "k_" prefix for global constants.
-* Use UPPER_CASE names for macros.
