+++ src/cleanup.hpp
@@ -18,7 +18,7 @@
 
 #pragma once
 
-#include "util.hpp"
+#include "Util.hpp"
 
 #include <string>
 
@@ -27,10 +27,10 @@ class Config;
 void clean_up_dir(const std::string& subdir,
                   uint64_t max_size,
                   uint32_t max_files,
-                  const util::ProgressReceiver& progress_receiver);
+                  const Util::ProgressReceiver& progress_receiver);
 
 void clean_up_all(const Config& config,
-                  const util::ProgressReceiver& progress_receiver);
+                  const Util::ProgressReceiver& progress_receiver);
 
 void wipe_all(const Config& config,
-              const util::ProgressReceiver& progress_receiver);
+              const Util::ProgressReceiver& progress_receiver);
